+++ languagetool-core/src/main/java/org/languagetool/tools/StringTools.java
@@ -24,7 +24,6 @@ import org.languagetool.Language;
 
 import java.io.*;
 import java.lang.Character;
-import java.util.Collection;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.regex.Pattern;
@@ -310,13 +309,6 @@ public final class StringTools {
       }
     }
     return sb.toString();
-  }
-
-  /**
-   * @deprecated use e.g. Java's {@code String.join()} (deprecated since 3.2)
-   */
-  public static String listToString(final Collection<String> l, final String delimiter) {
-    return String.join(delimiter, l);
   }
 
   /**